// apps/resolution-service/src/modules/penalty/domain/penalty.enums.ts
export enum PenaltyType {
  WARNING = "WARNING",
  FINE = "FINE",
  COMPENSATION = "COMPENSATION",
  SUSPEND = "SUSPEND",
  BAN = "BAN",
  REPUTATION_DEDUCTION = "REPUTATION_DEDUCTION",
  RESTRICT_ACTIVITY = "RESTRICT_ACTIVITY",
}

export enum PenaltyStatus {
  ISSUED = "ISSUED",
  EXECUTING = "EXECUTING",
  EXECUTED = "EXECUTED",
  EXECUTION_FAILED = "EXECUTION_FAILED",
  APPEALED = "APPEALED",
  UPHELD = "UPHELD",
  OVERTURNED = "OVERTURNED",
  WAIVED = "WAIVED",
}

export enum AppealStatus {
  SUBMITTED = "SUBMITTED",
  UNDER_REVIEW = "UNDER_REVIEW",
  UPHELD = "UPHELD",
  OVERTURNED = "OVERTURNED",
  REJECTED = "REJECTED",
}